Update firework.pde
All of the fireworks draw perfectly. The flare split of the fireworks is now always even. The functions for buttons have been designated. All that is left to do is draw each individual button, put text on/over the buttons, and make checks for where the mouse positions is to determine which button is clicked.