Update firework.pde
All coloration buttons are drawn and the loop button is one line away from being drawn. A divider has been set up to separate the buttons from the valid launch area. Click location checks have yet to be built, but they are the next and final element of the project to be implemented.